+++ packages/example-floors-astro/src/utils/types.ts
@@ -7,7 +7,7 @@ import type {
   TaggedTag,
 } from 'svgmapviewer-astro-floors'
 
-import { type XInfo } from '../../utils/schema'
+import { type XInfo } from './schema'
 
 declare module 'svgmapviewer' {
   interface Info {
